Commit for 2019.09.06.7z

+++ engine/inc/uf/utils/math/math.h
@@ -2,7 +2,7 @@
 
 namespace pod {
 	namespace Math {
-		typedef double num_t;
+		typedef float num_t;
 	}
 }
